# docs/swagger.yaml
basePath: /
...
host: localhost:3000
info:
  contact:
    email: support@cocobase.io
    name: Cocobase Support
  description: Backend as a Service with flexible collections and document management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cocobase API
  version: "1.0"
paths:
...
  /collections/{id}/realtime:
    get:
      description: WebSocket endpoint for real-time document changes. Send API key
        as first message.
      parameters:
      - description: Collection ID or name
        in: path
        name: id
        required: true
        type: string
      responses: {}
      summary: Subscribe to real-time collection updates
      tags:
      - Realtime
  /collections/{id}/realtime/stats:
    get:
      description: Get statistics about active WebSocket connections
      parameters:
      - description: Collection ID or name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get real-time connection statistics
      tags:
      - Realtime
...
  /realtime/stats:
    get:
      description: Get WebSocket connection statistics for all collections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all real-time statistics
      tags:
      - Realtime
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"